Extracted MqttTopicUtil from core
