+++ kura/org.eclipse.kura.ai.triton.server/META-INF/MANIFEST.MF
@@ -41,7 +41,6 @@ Import-Package: com.google.common.base;version="25.0.0",
  org.eclipse.kura.ai.inference;version="[1.0,1.1)",
  org.eclipse.kura.configuration;version="[1.1,2.0)",
  org.eclipse.kura.core.linux.executor;version="[1.0,2.0)",
- org.eclipse.kura.core.util;version="[1.0,2.0)",
  org.eclipse.kura.container.orchestration;version="[1.3,2.0)",
  org.eclipse.kura.crypto;version="[1.0,2.0)",
  org.eclipse.kura.executor;version="[1.0,2.0)",
