+++ kura/org.eclipse.kura.cloud.base.provider/META-INF/MANIFEST.MF
@@ -5,13 +5,14 @@ Bundle-SymbolicName: org.eclipse.kura.cloud.base.provider;singleton:=true
 Bundle-Version: 1.0.0.qualifier
 Bundle-Vendor: Eclipse Kura
 Require-Capability: osgi.ee;filter:="(&(osgi.ee=JavaSE)(version=1.8))"
-Export-Package: org.eclipse.kura.core.data;version="1.3.0"
+Export-Package: org.eclipse.kura.core.data;version="1.3.0",
+ org.eclipse.kura.core.data.util;version="1.0.0";x-internal:=true
 Service-Component: OSGI-INF/*.xml
 Bundle-ActivationPolicy: lazy
 Import-Package: org.eclipse.kura;version="[1.0,2.0)",
  org.eclipse.kura.configuration;version="[1.2,2.0)",
  org.eclipse.kura.connection.listener;version="[1.0,2.0)",
- org.eclipse.kura.core.util;version="[1.3,2.0)",
+ org.eclipse.kura.core.util;version="[2.0,3.0)",
  org.eclipse.kura.crypto;version="[1.3,2.0)",
  org.eclipse.kura.data;version="[1.1,2.0)",
  org.eclipse.kura.data.listener;version="[1.0,2.0)",
