+++ kura/org.eclipse.kura.cloud.base.provider/build.properties
@@ -1,5 +1,5 @@
 #
-#  Copyright (c) 2011, 2020 Eurotech and/or its affiliates and others
+#  Copyright (c) 2024 Eurotech and/or its affiliates and others
 #
 #  This program and the accompanying materials are made
 #  available under the terms of the Eclipse Public License 2.0
@@ -17,7 +17,6 @@ output.. = target/classes/
 bin.includes = META-INF/,\
                .,\
                OSGI-INF/,\
-               src/main/resources/,\
                about.html,\
                lib/org.eclipse.paho.client.mqttv3-1.2.1.k2.jar
 additional.bundles = slf4j.api,\
