+++ kura/org.eclipse.kura.cloud.base.provider/src/main/java/org/eclipse/kura/core/data/util/MqttTopicUtil.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2018, 2020 Eurotech and/or its affiliates and others
+ * Copyright (c) 2024 Eurotech and/or its affiliates and others
  * 
  * This program and the accompanying materials are made
  * available under the terms of the Eclipse Public License 2.0
@@ -10,7 +10,7 @@
  * Contributors:
  *  Eurotech
  *******************************************************************************/
-package org.eclipse.kura.core.util;
+package org.eclipse.kura.core.data.util;
 
 import org.eclipse.paho.client.mqttv3.MqttTopic;
