+++ kura/org.eclipse.kura.cloudconnection.eclipseiot.mqtt.provider/META-INF/MANIFEST.MF
@@ -24,7 +24,7 @@ Import-Package: com.eclipsesource.json;version="0.9.5",
  org.eclipse.kura.configuration;version="[1.1,2.0)",
  org.eclipse.kura.configuration.metatype;version="[1.1,2.0)",
  org.eclipse.kura.core.data;version="[1.0,2.0)",
- org.eclipse.kura.core.util;version="[1.0,2.0)",
+ org.eclipse.kura.core.util;version="[2.0,3.0)",
  org.eclipse.kura.data;version="[1.0,2.0)",
  org.eclipse.kura.data.listener;version="[1.0,1.1)",
  org.eclipse.kura.marshalling;version="[1.0,2.0)",
