+++ kura/org.eclipse.kura.cloudconnection.kapua.mqtt.provider/META-INF/MANIFEST.MF
@@ -27,8 +27,9 @@ Import-Package: com.eclipsesource.json;version="0.9.5",
  org.eclipse.kura.command;version="[1.2,1.3)",
  org.eclipse.kura.configuration;version="[1.1,2.0)",
  org.eclipse.kura.configuration.metatype;version="[1.1,2.0)",
- org.eclipse.kura.core.data,
- org.eclipse.kura.core.util;version="[1.0,2.0)",
+ org.eclipse.kura.core.data;version="[1.3,2.0)",
+ org.eclipse.kura.core.data.util;version="[1.0,2.0)",
+ org.eclipse.kura.core.util;version="[2.0,3.0)",
  org.eclipse.kura.crypto;version="[1.0,2.0)",
  org.eclipse.kura.data;version="[1.0,2.0)",
  org.eclipse.kura.data.listener;version="[1.0,1.1)",
