+++ kura/org.eclipse.kura.cloudconnection.raw.mqtt.provider/META-INF/MANIFEST.MF
@@ -17,7 +17,7 @@ Import-Package: org.eclipse.kura;version="[1.0,2.0)",
  org.eclipse.kura.cloudconnection.subscriber;version="[1.0,1.1)",
  org.eclipse.kura.cloudconnection.subscriber.listener;version="[1.0,2.0)",
  org.eclipse.kura.configuration;version="[1.1,2.0)",
- org.eclipse.kura.core.util;version="[1.2,2.0)",
+ org.eclipse.kura.core.data.util;version="[1.0,2.0)",
  org.eclipse.kura.data;version="[1.0,2.0)",
  org.eclipse.kura.data.listener;version="[1.0,1.1)",
  org.eclipse.kura.message;version="[1.0,2.0)",
