+++ kura/org.eclipse.kura.core.configuration/META-INF/MANIFEST.MF
@@ -23,7 +23,7 @@ Import-Package: com.eclipsesource.json;version="0.9.5",
  org.eclipse.kura.cloudconnection.request;version="[1.0,2.0)",
  org.eclipse.kura.configuration;version="[1.2,1.3)",
  org.eclipse.kura.configuration.metatype;version="[1.1,2.0)",
- org.eclipse.kura.core.util;version="[1.0,2.0)",
+ org.eclipse.kura.core.util;version="[2.0,3.0)",
  org.eclipse.kura.crypto;version="[1.0,2.0)",
  org.eclipse.kura.marshalling;version="[1.0,2.0)",
  org.eclipse.kura.message;version="[1.0,2.0)",
