+++ kura/org.eclipse.kura.core.net/META-INF/MANIFEST.MF
@@ -9,7 +9,6 @@ Bundle-ClassPath: .
 Bundle-ActivationPolicy: lazy
 Import-Package: org.eclipse.kura;version="[1.0,2.0)",
  org.eclipse.kura.configuration;version="[1.1,2.0)",
- org.eclipse.kura.core.util;version="[1.0,2.0)",
  org.eclipse.kura.executor;version="[1.0,2.0)",
  org.eclipse.kura.net;version="[2.0,3.0)",
  org.eclipse.kura.net.configuration;version="[1.1,2.0)",
