+++ kura/org.eclipse.kura.core/META-INF/MANIFEST.MF
@@ -8,7 +8,7 @@ Require-Capability: osgi.ee;filter:="(&(osgi.ee=JavaSE)(version=1.8))"
 Export-Package: org.eclipse.kura.core.linux.executor;version="1.0.0",
  org.eclipse.kura.core.linux.util;version="1.2.0",
  org.eclipse.kura.core.ssl;version="1.0.0",
- org.eclipse.kura.core.util;version="1.3.0";x-internal:=true
+ org.eclipse.kura.core.util;version="2.0.0";x-internal:=true
 Service-Component: OSGI-INF/*.xml
 Bundle-ActivationPolicy: lazy
 Import-Package: javax.crypto,
