+++ kura/org.eclipse.kura.core/build.properties
@@ -1,5 +1,5 @@
 #
-#  Copyright (c) 2011, 2020 Eurotech and/or its affiliates and others
+#  Copyright (c) 2011, 2024 Eurotech and/or its affiliates and others
 #
 #  This program and the accompanying materials are made
 #  available under the terms of the Eclipse Public License 2.0
@@ -17,9 +17,7 @@ output.. = target/classes/
 bin.includes = META-INF/,\
                .,\
                OSGI-INF/,\
-               src/main/resources/,\
-               about.html,\
-               lib/org.eclipse.paho.client.mqttv3-1.2.1.k2.jar
+               about.html
 additional.bundles = slf4j.api,\
                      org.eclipse.osgi,\
                      org.eclipse.equinox.io,\
