+++ kura/org.eclipse.kura.linux.clock/META-INF/MANIFEST.MF
@@ -16,7 +16,6 @@ Import-Package: com.google.gson;version="2.7.0",
  org.eclipse.kura.clock;version="[1.0,1.1)",
  org.eclipse.kura.configuration;version="[1.1,2.0)",
  org.eclipse.kura.core.linux.executor;version="[1.0,2.0)",
- org.eclipse.kura.core.util;version="[1.0,2.0)",
  org.eclipse.kura.crypto;version="[1.3,2.0)",
  org.eclipse.kura.executor;version="[1.0,2.0)",
  org.eclipse.kura.position;version="[1.0,2.0)",
