+++ kura/org.eclipse.kura.linux.command/META-INF/MANIFEST.MF
@@ -12,6 +12,5 @@ Import-Package: org.apache.commons.exec;version="[1.3,2.0)",
  org.eclipse.kura;version="[1.3,2.0)",
  org.eclipse.kura.command;version="[1.2,1.3)",
  org.eclipse.kura.core.linux.executor;version="[1.0,2.0)",
- org.eclipse.kura.core.util;version="[1.0,2.0)",
  org.eclipse.kura.executor;version="[1.0,2.0)",
  org.slf4j;version="1.6.4"
