+++ kura/org.eclipse.kura.linux.net/META-INF/MANIFEST.MF
@@ -22,7 +22,6 @@ Import-Package: javax.microedition.io,
  org.eclipse.kura.core.net;version="[1.0,2.0)",
  org.eclipse.kura.core.net.modem;version="[1.0,2.0)",
  org.eclipse.kura.core.net.util;version="[1.0,2.0)",
- org.eclipse.kura.core.util;version="[1.0,2.0)",
  org.eclipse.kura.executor;version="[1.0,2.0)",
  org.eclipse.kura.net;version="[2.0,3.0)",
  org.eclipse.kura.net.dhcp;version="[1.2,1.3)",
