+++ kura/org.eclipse.kura.nm/META-INF/MANIFEST.MF
@@ -22,7 +22,6 @@ Import-Package: javax.xml.bind;version="2.3.3",
  org.eclipse.kura.core.net;version="[1.0,2.0)",
  org.eclipse.kura.core.net.modem;version="[1.0,2.0)",
  org.eclipse.kura.core.net.util;version="[1.0,2.0)",
- org.eclipse.kura.core.util;version="[1.0,2.0)",
  org.eclipse.kura.crypto;version="[1.1,2.0)",
  org.eclipse.kura.executor;version="[1.0,2.0)",
  org.eclipse.kura.internal.linux.net.dns;version="[1.0,2.0)",
