+++ kura/org.eclipse.kura.web2/META-INF/MANIFEST.MF
@@ -47,7 +47,7 @@ Import-Package: com.eclipsesource.json;version="0.9.5",
  org.eclipse.kura.core.net;version="[1.0,2.0)",
  org.eclipse.kura.core.net.util;version="[1.0,2.0)",
  org.eclipse.kura.core.ssl;version="[1.0,2.0)",
- org.eclipse.kura.core.util;version="[1.0,2.0)",
+ org.eclipse.kura.core.util;version="[2.0,3.0)",
  org.eclipse.kura.crypto;version="[1.3,2.0)",
  org.eclipse.kura.data;version="[1.0,2.0)",
  org.eclipse.kura.db;version="[2.0,3.0)",
