+++ kura/test/org.eclipse.kura.core.testutil/META-INF/MANIFEST.MF
@@ -29,7 +29,7 @@ Import-Package: com.eclipsesource.json;version="0.9.5",
  org.eclipse.jetty.util.component;version="9.4.41",
  org.eclipse.kura;version="[1.6,2.0)",
  org.eclipse.kura.configuration;version="[1.2,2.0)",
- org.eclipse.kura.core.util;version="[1.3,2.0)",
+ org.eclipse.kura.core.data.util;version="[1.0,2.0)",
  org.eclipse.kura.crypto;version="[1.3,2.0)",
  org.eclipse.kura.data;version="[1.1,2.0)",
  org.eclipse.kura.data.transport.listener;version="[1.0,2.0)",
