+++ kura/test/org.eclipse.kura.linux.net.test/META-INF/MANIFEST.MF
@@ -8,7 +8,6 @@ Require-Capability: osgi.ee;filter:="(&(osgi.ee=JavaSE)(version=1.8))"
 Import-Package: org.eclipse.kura;version="[1.0,2.0)",
  org.eclipse.kura.core.net.util;version="[1.0,2.0)",
  org.eclipse.kura.core.testutil,
- org.eclipse.kura.core.util;version="[1.0,2.0)",
  org.eclipse.kura.executor;version="1.0.0",
  org.eclipse.kura.linux.net.dhcp;version="[2.0,3.0)",
  org.eclipse.kura.linux.net.iptables;version="[1.0,2.0)",
