+++ kura/test/org.eclipse.kura.rest.wire.provider.test/META-INF/MANIFEST.MF
@@ -15,7 +15,6 @@ Import-Package: com.eclipsesource.json;version="0.9.5",
  org.eclipse.kura.core.testutil.json;version="[1.0,2.0)",
  org.eclipse.kura.core.testutil.requesthandler;version="[1.0,2.0)",
  org.eclipse.kura.core.testutil.service;version="[1.0,2.0)",
- org.eclipse.kura.core.util;version="[1.3,2.0)",
  org.eclipse.kura.crypto;version="[1.3,2.0)",
  org.eclipse.kura.data;version="[1.1,2.0)",
  org.eclipse.kura.data.transport.listener;version="[1.0,2.0)",
